Lots of work put into porting to Opengl 3.3+. Got it back, more or less, to its previously functional state.

+++ Exort/src/com/doobs/exort/util/font/CharInfoLoader.java
@@ -3,7 +3,7 @@ package com.doobs.exort.util.font;
 import java.io.*;
 import java.util.*;
 
-import com.doobs.exort.util.texture.*;
+import com.doobs.modern.util.texture.*;
 
 public class CharInfoLoader {
 
@@ -23,7 +23,7 @@ public class CharInfoLoader {
 					temp = new double[8];
 
 					for (int i = 0; i < line.length(); i++) {
-						if ((int) line.charAt(i) >= 48 && (int) line.charAt(i) <= 57 || line.charAt(i) == '.')
+						if (line.charAt(i) >= 48 && line.charAt(i) <= 57 || line.charAt(i) == '.')
 							number += line.charAt(i);
 						else if (!number.equals("")) {
 							temp[numbers] = Double.valueOf(number);
